@@ -14,12 +14,13 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package com.ns.yc.ycstatelib;
+package com.yc.ycstatelib;
 
 import android.content.Context;
-import android.support.annotation.LayoutRes;
 import android.view.View;
 import android.view.ViewStub;
+
+import androidx.annotation.LayoutRes;
 
 
 /**
