@@ -1,9 +1,6 @@
 package com.pedaily.yc.ycstatelayout.adapter;
 
-/**
- * @author billy.qi
- * @since 19/3/21 10:03
- */
+
 public interface Constants {
 
     String HIDE_LOADING_STATUS_MSG = "hide_loading_status_msg";
