@@ -2,13 +2,14 @@ package com.pedaily.yc.ycstatelayout.base;
 
 import android.graphics.Color;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
 import android.widget.LinearLayout;
 
-import com.ns.yc.ycstatelib.StateLayoutManager;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
+import com.yc.ycstatelib.StateLayoutManager;
 import com.pedaily.yc.ycstatelayout.R;
 
 
