@@ -2,10 +2,11 @@ package com.pedaily.yc.ycstatelayout.base;
 
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.annotation.UiThread;
-import android.support.v4.app.Fragment;
 import android.util.Log;
+
+import androidx.annotation.Nullable;
+import androidx.annotation.UiThread;
+import androidx.fragment.app.Fragment;
 
 /**
  * <pre>
