@@ -3,15 +3,16 @@ package com.pedaily.yc.ycstatelayout.base;
 import android.app.Activity;
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.FragmentTransaction;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
 
-import com.ns.yc.ycstatelib.StateLayoutManager;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.FragmentTransaction;
+
+import com.yc.ycstatelib.StateLayoutManager;
 import com.pedaily.yc.ycstatelayout.R;
 
 
